requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Cerberus==1.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.7
Django==4.2.8
djangorestframework==3.14.0
docopt==0.6.2
idna==3.6
isort==5.13.0
pep517==0.13.1
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requirementslib==3.0.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
yarg==0.1.9
